+++ Packs/ExportIndicators/Integrations/ExportIndicators/ExportIndicators.py
@@ -902,7 +902,7 @@ def update_outbound_command(args, params):
 
     query = args.get('query')
     # in case no query is entered take the query in the integration params
-    if len(query) == 0:
+    if not query:
         query = params.get('indicators_query')
 
     out_format = args.get('format')