+++ Packs/Base/Scripts/CommonServerPython/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-
+  - Added retry mechanism to the BaseClient.
+  - Fixed an issue where the **appendContext** function did not behave as expected.
 
 ## [20.4.1] - 2020-04-29
   - Deprecated the following enums: